requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
certifi==2017.11.5
chardet==3.0.4
click==6.7
configparser==3.5.0
enum34==1.1.6
Flask==0.12.2
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pylint==1.7.4
python-google-places==1.4.1
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.11
